Stable, refactored solution - possible performance issues
This commit introduces the following changes:
* Stable, refactored implementation of a solution to
  CodinGame's Tan Network puzzle, found under src/
* Unit and functionality tests, found under test/
* Formatting script for CodinGame's online IDE, found under tools/
* Minimal documentation (docstrings and readme file)

Limitations:

This solutions passes all the regular CodinGame's test cases,
but fails 1 of the 6 'submit' test cases.

As the former does not provide any information the failure, one can only
make assumptions about the failure's root cause.

Since this failing test case ensures that the programs correctly finds
a long path, and that the similar regular test case is passing, the
program's performance could be a cause of failure.